Fix Tracker projects layout
